Ecrad class in the python binding
